requirements.txt
asgiref==3.5.1
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
mysqlclient==2.1.0
pycodestyle==2.8.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
zipp==3.8.0
